+++ src/service/history/HistoryService.js
@@ -1,24 +1,25 @@
+import * as _ from 'fxjs';
+import * as L from 'fxjs/Lazy';
 import {moduleLogger} from '@src/logger';
 import {objectToString} from '@src/util/conversion';
-import {sequelize} from '@src/database/Sequelize';
 import * as HistoryRepository from '@src/repository/HistoryRepository';
 import * as HistoryAndTagRepository from '@src/repository/HistoryAndTagRepository';
+import {
+    NotFoundException,
+    UnauthorizedException,
+} from '@src/exception/CommonException';
 
 const logger = moduleLogger('HistoryService');
 
-export const postHistory = async (historyInfo) => {
-    logger.debug(`postHistory start, { "historyInfo": ${objectToString(historyInfo)} }`);
+export const postHistory = async ({playTime, accountId, incenseId, tagIds}) => {
+    logger.debug(`postHistory start playTime: ${playTime}, accountId: ${accountId}, incenseId: ${incenseId}`);
 
     const savedHistory = await HistoryRepository.saveHistory({
-        playTime: historyInfo.playTime,
-        accountId: historyInfo.accountId,
-        incenseId: historyInfo.incenseId,
-        createdAt: sequelize.literal('CURRENT_TIMESTAMP'),
+        playTime, accountId, incenseId,
     });
 
     logger.info(`save history success, { "historyInfo": ${objectToString(savedHistory)} }`);
 
-    const tagIds = historyInfo.tagIds;
     tagIds.map(async (tagId) => {
         await HistoryAndTagRepository.saveHistoryAndTag({
             historyId: savedHistory.id,
@@ -45,7 +46,7 @@ export const postImage = async ({historyId, imageUrl}) => {
 export const getHistory = async (accountId) => {
     logger.debug(`getHistory start, { "accountId": ${accountId} }`);
 
-    const histories = await HistoryRepository.getHistory(accountId);
+    const histories = await HistoryRepository.getHistories(accountId);
 
     if (!histories) {
         logger.error(`cannot find history info with accountId, { "accountId": "${accountId}" }`);
@@ -57,10 +58,31 @@ export const getHistory = async (accountId) => {
     return histories;
 };
 
-export const deleteHistory = async (historyId) => {
+export const deleteHistory = async ({historyId, accountId}) => {
     logger.debug(`deleteHistory start, { "historyId": ${objectToString(historyId)} }`);
+
+    const history = await HistoryRepository.findOneById(historyId);
+    if (history.accountId !== accountId) {
+        logger.error(`this account cannot delete this history, { "accountId": "${accountId}", "historyAccountId": "${history.accountId}" }`);
+        throw new UnauthorizedException();
+    }
 
     await HistoryRepository.deleteHistory(historyId);
 
     logger.info(`deleteHistory success}`);
+};
+
+export const getReportsData = async ({accountId}) => {
+    const reportsData = await HistoryRepository.findAllByAccountId(accountId);
+
+    const reportsResult = _.go(reportsData,
+        _.groupBy((report) => report.incenseId),
+        L.entries,
+        L.map(([key, value]) => [key, value.length]),
+        _.object,
+    );
+
+    logger.info(`getReportsData success, reports: ${reportsResult}`);
+
+    return reportsResult;
 };
